Stricter compiler options (#25)
* use conan an its python_require feature to populate compiler flags

* distinguish debug and release flags

* missing file

* reorganize options, apply review comments

* adjust conn conanfile

* lto and cxx standard in conan file

* make lto a conan option defaulting to true

* ctrl memstore shared crypto neb libs

* format code

* fix gcc errors in crash-consensus

* export conanfiles under conan/exports/* in build.py

* Revert "fix gcc errors in crash-consensus"

This reverts commit d924f80d83d5ecde13e5b6e41c9905f9fd5ce103.

* disable crash-consensus in makefile from building

* fix clang sign-conversion

* memoryrights instead of int

* remove leader-election, clang-format protobuf, add cppstd in profile and apply review comments

+++ conn/src/exchanger.cpp
@@ -102,7 +102,7 @@ std::pair<bool, int> ConnectionExchanger::valid_ids() const {
     return std::make_pair(false, 0);
   }
 
-  if (double(max) > gapFactor * (remote_ids.size() + 1)) {
+  if (double(max) > gapFactor * static_cast<double>((remote_ids.size() + 1))) {
     return std::make_pair(false, 0);
   }
