+++ conn/src/rc.hpp
@@ -107,7 +107,7 @@ class ReliableConnection {
   void connect(RemoteConnection &rci);
   void reconnect();
 
-  bool postSendSingle(RdmaReq req, uint64_t req_id, void *buf, uint64_t len,
+  bool postSendSingle(RdmaReq req, uint64_t req_id, void *buf, uint32_t len,
                       uintptr_t remote_addr);
 
   // Only re-use this method when the previous WR posted by this method is
@@ -116,9 +116,9 @@ class ReliableConnection {
   // (which is the case when the length of the payload is smaller or equal to
   // `MaxInlining`) one can reuse this method right after it returns.
   bool postSendSingleCached(RdmaReq req, uint64_t req_id, void *buf,
-                            uint64_t len, uintptr_t remote_addr);
+                            uint32_t len, uintptr_t remote_addr);
 
-  bool postSendSingle(RdmaReq req, uint64_t req_id, void *buf, uint64_t len,
+  bool postSendSingle(RdmaReq req, uint64_t req_id, void *buf, uint32_t len,
                       uint32_t lkey, uintptr_t remote_addr);
 
   bool pollCqIsOK(CQ cq, std::vector<struct ibv_wc> &entries);
